+++ tiledb/sm/c_api/tiledb.cc
@@ -1573,6 +1573,34 @@ int32_t tiledb_attribute_dump(
   return TILEDB_OK;
 }
 
+int32_t tiledb_attribute_set_fill_value(
+    tiledb_ctx_t* ctx,
+    tiledb_attribute_t* attr,
+    const void* value,
+    uint64_t size) {
+  if (sanity_check(ctx) == TILEDB_ERR || sanity_check(ctx, attr) == TILEDB_ERR)
+    return TILEDB_ERR;
+
+  if (SAVE_ERROR_CATCH(ctx, attr->attr_->set_fill_value(value, size)))
+    return TILEDB_ERR;
+
+  return TILEDB_OK;
+}
+
+int32_t tiledb_attribute_get_fill_value(
+    tiledb_ctx_t* ctx,
+    tiledb_attribute_t* attr,
+    const void** value,
+    uint64_t* size) {
+  if (sanity_check(ctx) == TILEDB_ERR || sanity_check(ctx, attr) == TILEDB_ERR)
+    return TILEDB_ERR;
+
+  if (SAVE_ERROR_CATCH(ctx, attr->attr_->get_fill_value(value, size)))
+    return TILEDB_ERR;
+
+  return TILEDB_OK;
+}
+
 /* ********************************* */
 /*              DOMAIN               */
 /* ********************************* */